Gold coins move towards player, player character screen
Started on a new GUI element that displays the player's stats, along with the amount of money they have. Also made it so that gold coins do not get added to the inventory, and will instead increment a value. Furthermore, they also move towards the player when they are close enough.

+++ Assets/Scripts/UI/PlayerUi.cs
@@ -112,7 +112,7 @@ namespace UI
         private void UpdateManaBar()
         {
             float prevManaFill    = manaBarFill.fillAmount;
-            float currentManaFill = (float)_playerAbilitySystem.CurrentMana / _playerAbilitySystem.maxMana;
+            float currentManaFill = _playerAbilitySystem.CurrentMana / _playerAbilitySystem.maxMana;
             if (currentManaFill > prevManaFill) prevManaFill = Mathf.Min(prevManaFill + fillSmoothness, currentManaFill);
             else if (currentManaFill < prevManaFill)
                 prevManaFill = Mathf.Max(prevManaFill - fillSmoothness, currentManaFill);
